requirements.txt
elasticsearch==7.5.1
mysql-replication==0.21
urllib3==1.25.8
chardet==3.0.4
PyMySQL==0.9.3
unittest2==1.1.0